Add scripts create alternative database for the project
simulate a scene where some objects (spheres for example) move in the room, while we record their acceleration in 3 axis and also a video from a "ceiling camera". 
we control their movement (position, velocity and acceleration) and the camera configurations, and get a simulation instead of the MatchNMingle dataset we can use to test the alternative diffusion algorithm.